requirements.txt
bleach==6.1.0
bcrypt
fastapi==0.104.1
passlib==1.7.4
pydantic==1.10.13
python-dotenv==1.0.0
python_jose==3.3.0
sqlmodel==0.0.12
uvicorn
psycopg2-binary
python-multipart
pytz